api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
anthropic==0.40.0
apscheduler==3.10.4
